requirements.txt
torch
transformers
sentence-transformers
langchain
faiss-cpu
PyPDF2
streamlit
ctransformers
unstructured
streamlit
pypdf
